refactor: drop app/servers/service.py shim by migrating last imports (Resolves #276)
PR #275 left `app/servers/service.py` as a 41-line re-export shim because
unit tests still imported the legacy module path. PR #290 already deleted
`app/services/server_service.py` (the second shim referenced in this issue).

After auditing the codebase, no caller actually used the
`patch("app.servers.service.X")` form the shim's docstring justified, and
only five remaining sources imported from `app.servers.service`. They are
migrated to the canonical paths:

- `ServerJarService`, `ServerService`, `ServerDatabaseService`,
  `ServerFileSystemService`, `ServerSecurityValidator` ->
  `app.servers.application.service`
- `ServerValidationService` -> `app.servers.adapters._legacy_helpers`

With every consumer migrated, the shim is removed outright (43 -> 0 lines,
better than the 8-line target). The docstring on
`app/servers/application/service.py` is updated to record the deletion.

Migrated imports:
- tests/integration/test_database_version_integration.py (1)
- tests/unit/servers/test_service.py (1 grouped import covering 6 symbols)
- tests/test_security.py (3 imports: top-level + two inline within tests)

Test plan:
- uv run ruff check on touched files (only pre-existing F841 noise unrelated to this change)
- uv run pytest tests/unit/servers/test_service.py tests/test_security.py
  tests/integration/test_database_version_integration.py -m "not slow" -> 83 passed
- uv run pytest tests/unit -m "not slow" -> all green
- uv run pytest tests/integration -m "not slow" -> all green
